Splitt opp grunnlag og vilkårperiode i to ulike servicer for å lage mindre innhold i en service
