+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vilkår/vilkårperiode/VilkårperiodeService.kt
@@ -1,6 +1,5 @@
 package no.nav.tilleggsstonader.sak.vilkår.vilkårperiode
 
-import no.nav.tilleggsstonader.kontrakter.ytelse.EnsligForsørgerStønadstype
 import no.nav.tilleggsstonader.sak.behandling.BehandlingService
 import no.nav.tilleggsstonader.sak.behandling.BehandlingUtil.validerBehandlingIdErLik
 import no.nav.tilleggsstonader.sak.behandling.domain.Saksbehandling
@@ -10,11 +9,6 @@ import no.nav.tilleggsstonader.sak.infrastruktur.database.repository.findByIdOrT
 import no.nav.tilleggsstonader.sak.infrastruktur.exception.brukerfeilHvis
 import no.nav.tilleggsstonader.sak.infrastruktur.exception.brukerfeilHvisIkke
 import no.nav.tilleggsstonader.sak.infrastruktur.exception.feilHvis
-import no.nav.tilleggsstonader.sak.infrastruktur.sikkerhet.BehandlerRolle
-import no.nav.tilleggsstonader.sak.opplysninger.aktivitet.RegisterAktivitetService
-import no.nav.tilleggsstonader.sak.opplysninger.søknad.SøknadService
-import no.nav.tilleggsstonader.sak.opplysninger.ytelse.YtelseService
-import no.nav.tilleggsstonader.sak.tilgang.TilgangService
 import no.nav.tilleggsstonader.sak.vilkår.stønadsperiode.StønadsperiodeValideringUtil
 import no.nav.tilleggsstonader.sak.vilkår.stønadsperiode.domain.StønadsperiodeRepository
 import no.nav.tilleggsstonader.sak.vilkår.stønadsperiode.dto.tilSortertDto
@@ -42,24 +36,11 @@ import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.dto.VilkårperioderDto
 import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.dto.VilkårperioderResponse
 import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.dto.tilDto
 import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.felles.Vilkårstatus
-import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.grunnlag.GrunnlagAktivitet
-import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.grunnlag.GrunnlagYtelse
-import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.grunnlag.HentetInformasjon
-import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.grunnlag.PeriodeGrunnlagYtelse
-import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.grunnlag.RegisterAktivitet
-import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.grunnlag.SlåSammenPeriodeGrunnlagYtelseUtil.slåSammenOverlappendeEllerPåfølgende
-import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.grunnlag.VilkårperioderGrunnlag
-import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.grunnlag.VilkårperioderGrunnlagDomain
 import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.grunnlag.VilkårperioderGrunnlagRepository
-import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.grunnlag.tilDomenetype
 import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.grunnlag.tilDto
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
-import java.time.LocalDate
-import java.time.LocalDateTime
-import java.time.YearMonth
-import java.time.temporal.ChronoUnit
 import java.util.UUID
 
 @Service
@@ -68,10 +49,7 @@ class VilkårperiodeService(
     private val vilkårperiodeRepository: VilkårperiodeRepository,
     private val stønadsperiodeRepository: StønadsperiodeRepository,
     private val vilkårperioderGrunnlagRepository: VilkårperioderGrunnlagRepository,
-    private val registerAktivitetService: RegisterAktivitetService,
-    private val ytelseService: YtelseService,
-    private val søknadService: SøknadService,
-    private val tilgangService: TilgangService,
+    private val vilkårperiodeGrunnlagService: VilkårperiodeGrunnlagService,
 ) {
 
     private val logger = LoggerFactory.getLogger(javaClass)
@@ -86,132 +64,11 @@ class VilkårperiodeService(
     }
 
     fun hentVilkårperioderResponse(behandlingId: BehandlingId): VilkårperioderResponse {
-        val grunnlagsdataVilkårsperioder = hentEllerOpprettGrunnlag(behandlingId)
+        val grunnlagsdataVilkårsperioder = vilkårperiodeGrunnlagService.hentEllerOpprettGrunnlag(behandlingId)
 
         return VilkårperioderResponse(
             vilkårperioder = hentVilkårperioderDto(behandlingId),
             grunnlag = grunnlagsdataVilkårsperioder?.tilDto(),
-        )
-    }
-
-    @Transactional
-    fun oppdaterGrunnlag(behandlingId: BehandlingId) {
-        val behandling = behandlingService.hentBehandling(behandlingId)
-        feilHvis(behandling.status.behandlingErLåstForVidereRedigering()) {
-            "Kan ikke oppdatere grunnlag når behandlingen er låst"
-        }
-        feilHvis(behandling.steg != StegType.INNGANGSVILKÅR) {
-            "Kan ikke oppdatere grunnlag når behandlingen er i annet steg enn vilkår."
-        }
-
-        val eksisterendeGrunnlag = vilkårperioderGrunnlagRepository.findByIdOrThrow(behandlingId)
-
-        val eksisterendeHentetFom = eksisterendeGrunnlag.grunnlag.hentetInformasjon.fom
-        val tom = YearMonth.now().plusYears(1).atEndOfMonth()
-
-        val nyGrunnlagsdata = hentGrunnlagsdata(behandlingId, eksisterendeHentetFom, tom)
-        vilkårperioderGrunnlagRepository.update(eksisterendeGrunnlag.copy(grunnlag = nyGrunnlagsdata))
-        val tidSidenForrigeHenting =
-            ChronoUnit.HOURS.between(nyGrunnlagsdata.hentetInformasjon.tidspunktHentet, LocalDateTime.now())
-        logger.info("Oppdatert grunnlagsdata for behandling=$behandlingId timerSidenForrige=$tidSidenForrigeHenting")
-    }
-
-    private fun hentEllerOpprettGrunnlag(behandlingId: BehandlingId): VilkårperioderGrunnlag? {
-        val grunnlag = vilkårperioderGrunnlagRepository.findByBehandlingId(behandlingId)?.grunnlag
-
-        return if (grunnlag != null) {
-            grunnlag
-        } else if (behandlingErLåstForVidereRedigering(behandlingId)) {
-            null
-        } else {
-            opprettGrunnlagsdata(behandlingId).grunnlag
-        }
-    }
-
-    private fun opprettGrunnlagsdata(behandlingId: BehandlingId): VilkårperioderGrunnlagDomain {
-        brukerfeilHvisIkke(tilgangService.harTilgangTilRolle(BehandlerRolle.SAKSBEHANDLER)) {
-            "Behandlingen er ikke påbegynt. Kan ikke opprette vilkårperiode hvis man ikke er saksbehandler"
-        }
-
-        val søknadMetadata = søknadService.hentSøknadMetadata(behandlingId)
-        val utgangspunktDato = søknadMetadata?.mottattTidspunkt?.toLocalDate() ?: LocalDate.now()
-
-        val fom = YearMonth.from(utgangspunktDato).minusMonths(3).atDay(1)
-        val tom = YearMonth.from(utgangspunktDato).plusYears(1).atEndOfMonth()
-
-        val grunnlag = hentGrunnlagsdata(behandlingId, fom, tom)
-        return vilkårperioderGrunnlagRepository.insert(
-            VilkårperioderGrunnlagDomain(
-                behandlingId = behandlingId,
-                grunnlag = grunnlag,
-            ),
-        )
-    }
-
-    private fun hentGrunnlagsdata(
-        behandlingId: BehandlingId,
-        fom: LocalDate,
-        tom: LocalDate,
-    ): VilkårperioderGrunnlag {
-        return VilkårperioderGrunnlag(
-            aktivitet = hentGrunnlagAktvititet(behandlingId, fom, tom),
-            ytelse = hentGrunnlagYtelse(behandlingId, fom, tom),
-            hentetInformasjon = HentetInformasjon(
-                fom = fom,
-                tom = tom,
-                tidspunktHentet = LocalDateTime.now(),
-            ),
-        )
-    }
-
-    private fun hentGrunnlagAktvititet(
-        behandlingId: BehandlingId,
-        fom: LocalDate,
-        tom: LocalDate,
-    ) = GrunnlagAktivitet(
-        aktiviteter = registerAktivitetService.hentAktiviteterForGrunnlagsdata(
-            ident = behandlingService.hentSaksbehandling(behandlingId).ident,
-            fom = fom,
-            tom = tom,
-        ).map {
-            RegisterAktivitet(
-                id = it.id,
-                fom = it.fom,
-                tom = it.tom,
-                type = it.type,
-                typeNavn = it.typeNavn,
-                status = it.status,
-                statusArena = it.statusArena,
-                antallDagerPerUke = it.antallDagerPerUke,
-                prosentDeltakelse = it.prosentDeltakelse,
-                erStønadsberettiget = it.erStønadsberettiget,
-                erUtdanning = it.erUtdanning,
-                arrangør = it.arrangør,
-                kilde = it.kilde,
-            )
-        },
-    )
-
-    private fun hentGrunnlagYtelse(
-        behandlingId: BehandlingId,
-        fom: LocalDate,
-        tom: LocalDate,
-    ): GrunnlagYtelse {
-        val ytelserFraRegister = ytelseService.hentYtelseForGrunnlag(behandlingId = behandlingId, fom = fom, tom = tom)
-
-        return GrunnlagYtelse(
-            perioder = ytelserFraRegister.perioder
-                .filter { it.aapErFerdigAvklart != true }
-                .filter { it.ensligForsørgerStønadstype != EnsligForsørgerStønadstype.BARNETILSYN }
-                .map {
-                    PeriodeGrunnlagYtelse(
-                        type = it.type,
-                        fom = it.fom,
-                        tom = it.tom,
-                        ensligForsørgerStønadstype = it.ensligForsørgerStønadstype?.tilDomenetype(),
-                    )
-                }
-                .slåSammenOverlappendeEllerPåfølgende(),
         )
     }
 
@@ -350,7 +207,4 @@ class VilkårperiodeService(
         val kopiertePerioderMedReferanse = eksisterendeVilkårperioder.map { it.kopierTilBehandling(nyBehandlingId) }
         vilkårperiodeRepository.insertAll(kopiertePerioderMedReferanse)
     }
-
-    private fun behandlingErLåstForVidereRedigering(behandlingId: BehandlingId) =
-        behandlingService.hentBehandling(behandlingId).status.behandlingErLåstForVidereRedigering()
 }
